feat(avo-1870): splice some features into separate files
