+++ src/admin/shared/components/FilterTable/FilterTable.utils.ts
@@ -1,0 +1,13 @@
+import { isArray, isEmpty, isNil, isPlainObject, isString, omitBy } from 'lodash';
+
+// Removes all props where the value is undefined, null, [], {}, ''
+export function cleanupObject(obj: any) {
+	return omitBy(
+		obj,
+		(value: any) =>
+			isNil(value) ||
+			(isString(value) && !value.length) ||
+			((isPlainObject(value) || isArray(value)) && isEmpty(value)) ||
+			(isPlainObject(value) && value.gte === '' && value.lte === '')
+	);
+}
